Move DuplicateCache checking to workers
